Remove full version of Lena
With torch real spreading use, a few colleagues and I were thinking that
it may be time to remove lena_full -- it's hard to find a good reason
for it to be there.

I admit that I found it fun when it was first shown to me, like a fun wink.
But actually, thinking back, and while I'm not the most Politically Correct
guy (oh boy...), even I am now uncomfortable towards my female colleagues.
We talk a lot about diversity and making machine learning less of a
male-dominated field, let's act civilized and make them more welcome.
Yeah, it was fun for a while to feel like brogrammers, but now we're grown ups ;-)